TryingToFixEditEvent

+++ ClientApp/src/app/admin-crud-event/admin-crud-event.component.ts
@@ -1,9 +1,9 @@
 import { HttpClient } from '@angular/common/http';
-import { Inject } from '@angular/core';
-import { Component, OnInit } from '@angular/core';
-import { delay } from 'rxjs/internal/operators/delay';
+import { Component, Inject, OnInit } from '@angular/core';
+import { MatSnackBar } from '@angular/material';
 import { DataService } from '../data.service';
 import { Event } from '../Event'
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-admin-crud-event',
@@ -13,6 +13,13 @@ import { Event } from '../Event'
 export class AdminCrudEventComponent implements OnInit {
   testEvent: Event;
   eventToEdit: Event[];
+  confirmCreateEvent: boolean;
+  confirmDeleteEvent: boolean;
+  confirmEditEvent: boolean;
+  createButton: boolean;
+  editButton: boolean;
+  deleteButton: boolean;
+  cancelButton: boolean;
 
   eventTitle: String;
   eventLink: String;
@@ -27,7 +34,7 @@ export class AdminCrudEventComponent implements OnInit {
   eventDelete: String;
   event: Event[];
 
-  constructor(private http: HttpClient, @Inject('BASE_URL') private baseUrl: string,private data: DataService) {
+  constructor(private dialog: MatDialog, private _snackBar: MatSnackBar, private http: HttpClient, @Inject('BASE_URL') private baseUrl: string,private data: DataService) {
 
     this.http.get<Event[]>(this.baseUrl + 'api/Events').subscribe(result => {
       this.event = result;
@@ -37,11 +44,53 @@ export class AdminCrudEventComponent implements OnInit {
   }
 
   ngOnInit() {
+    this.createButton = true;
+    this.editButton = false;
+    this.deleteButton = false;
+    this.cancelButton = false;
+  }
+
+  successSnackBar(message: string) {
+    this._snackBar.open(message, "Close", {duration: 5000})
+    }
+
+  failureSnackBar(message: string, action: string) {
+    this._snackBar.open(message, action);
   }
 
   deleteEvent(){
+    this.createButton = false;
+    this.editButton = false;
+    this.deleteButton = true;
+    this.cancelButton = true;
     console.log(this.eventDelete);
-    this.data.deleteEvent(this.eventDelete);
+
+  }
+
+  confirmDelete(){
+    this.createButton = true;
+    this.editButton = false;
+    this.deleteButton = false;
+    this.cancelButton = false;
+    this.openDeleteDialog();
+    if(this.confirmDeleteEvent == true){
+      this.data.deleteEvent(this.eventDelete);
+      this.successSnackBar("Event Deleted!");
+    }
+
+  }
+
+  openDeleteDialog(){
+    const dialogRef = this.dialog.open(DeleteEventDialog);
+    dialogRef.afterClosed().subscribe(result => {
+      if(result == true){
+        this.confirmDeleteEvent = true;
+      }
+      else{
+        this.confirmDeleteEvent = false;
+      }
+      console.log(`Dialog result: ${result}`);
+    })
   }
 
   resetEventName(){
@@ -64,8 +113,36 @@ export class AdminCrudEventComponent implements OnInit {
       googleCalID: "abc123"
     }
 
-    this.data.postEvent(event);
+    if(event.bandName != null && event.eventStart != null &&
+      event.eventEnd != null && event.maxNumberOfSeats != null &&
+      event.ticketPrice != null && event.refundCutOffDate != null){
 
+        this.openCreateDialog();
+
+        if(this.confirmCreateEvent == true){
+          this.data.postEvent(event);
+          console.log("Creating Event");
+          this.successSnackBar("Event Created!");
+        }
+
+      }
+    else{
+      this.failureSnackBar("Please verify all fields with a red '*' are filled out accurately.", "Close");
+    }
+
+  }
+
+  openCreateDialog(){
+    const dialogRef = this.dialog.open(CreateEventDialog);
+    dialogRef.afterClosed().subscribe(result => {
+      if(result == true){
+        this.confirmCreateEvent = true;
+      }
+      else{
+        this.confirmCreateEvent = false;
+      }
+      console.log(`Dialog result: ${result}`);
+    })
   }
 
   filterEvents(){
@@ -83,11 +160,51 @@ export class AdminCrudEventComponent implements OnInit {
   }
 
   editEvent(){
+
+    this.createButton = false;
+    this.editButton = true;
+    this.deleteButton = false;
+    this.cancelButton = true;
+
     console.log("Edit Event");
-    this.data.getSpecificEvent(this.eventDelete);
+    //this.data.getSpecificEvent(this.eventDelete);
+
+    //this.eventTitle = this.data.getSpecificEvent(this.eventDelete);
 
     console.log(this.data.eventTitle);
     this.setFieldsForEdit();
+  }
+
+  confirmEdit(){
+    this.createButton = true;
+    this.editButton = false;
+    this.deleteButton = false;
+    this.cancelButton = false;
+    this.openEditDialog();
+    if(this.confirmDeleteEvent == true){
+      //this.data.editEvent(this.eventDelete);
+      this.successSnackBar("Event Edited!");
+    }
+  }
+
+  openEditDialog(){
+    const dialogRef = this.dialog.open(EditEventDialog);
+    dialogRef.afterClosed().subscribe(result => {
+      if(result == true){
+        this.confirmDeleteEvent = true;
+      }
+      else{
+        this.confirmDeleteEvent = false;
+      }
+      console.log(`Dialog result: ${result}`);
+    })
+  }
+
+  cancelEditDelete(){
+    this.createButton = true;
+    this.editButton = false;
+    this.deleteButton = false;
+    this.cancelButton = false;
   }
 
   setFieldsForEdit(){
@@ -95,3 +212,21 @@ export class AdminCrudEventComponent implements OnInit {
   }
 
 }
+
+@Component({
+  selector: 'create-event-dialog',
+  templateUrl: './create-event-dialog.html',
+})
+export class CreateEventDialog{}
+
+@Component({
+  selector: 'delete-event-dialog',
+  templateUrl: './delete-event-dialog.html',
+})
+export class DeleteEventDialog{}
+
+@Component({
+  selector: 'edit-event-dialog',
+  templateUrl: './edit-event-dialog.html',
+})
+export class EditEventDialog{}
